perf(bytecode): reduce instruction size to 4 bytes
feat(compiler): constants are no longer duplicated in their respective Vecs

feat(compiler): add nice error message when max number of constants is exceeded

+++ src/eval/call_frame.rs
@@ -33,11 +33,11 @@ impl CallFrame {
         instr
     }
 
-    pub fn rel_jmp(&mut self, offset: u32) {
+    pub fn rel_jmp(&mut self, offset: u16) {
         self.ip += offset as usize;
     }
 
-    pub fn rel_loop(&mut self, offset: u32) {
+    pub fn rel_loop(&mut self, offset: u16) {
         self.ip -= offset as usize;
     }
